+++ src/Controller/InstituicoesController.php
@@ -146,13 +146,14 @@ class InstituicoesController extends AppController
             $this->Flash->error(__('Instituição não encontrada.'));
             return $this->redirect(['action' => 'index']);
         }
-        $this->request->allowMethod(['post', 'delete']);
-        $this->Authorization->authorize($instituicao);
-        if ($this->Instituicoes->delete($instituicao)) {
-            $this->Flash->success(__('Instituição de estágio excluída.'));
-        } else {
+        if ($this->request->is(['post', 'delete'])) {
+            $this->Authorization->authorize($instituicao);
+            if ($this->Instituicoes->delete($instituicao)) {
+                $this->Flash->success(__('Instituição de estágio excluída.'));
+            } else {
             $this->Flash->error(__('Instituição de estágio não foi excluída.'));
             return $this->redirect(['action' => 'view', $instituicao->id]);
+            }
         }
         return $this->redirect(['action' => 'index']);
     }
